refactor: Refatora o controller para incluir o responseEntity nas minhas rotas e customizar a resposta das requisições do cliente

+++ src/main/java/dev/java10x/cadastrodeninjas/Ninjas/NinjaController.java
@@ -1,5 +1,7 @@
 package dev.java10x.cadastrodeninjas.Ninjas;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,31 +17,55 @@ public class NinjaController {
     }
 
     @GetMapping("/boasvindas")
-    public String boasVindas() { return  "Boas vindas nessa rota"; }
+    public String boasVindas() {
+        return "Boas vindas nessa rota";
+    }
 
     @PostMapping("/criar")
-    public NinjaDTO criarNinja(@RequestBody NinjaDTO ninja) {
-        return ninjaService.criarNinja(ninja);
+    public ResponseEntity<String> criarNinja(@RequestBody NinjaDTO ninja) {
+        NinjaDTO ninjaDTO = ninjaService.criarNinja(ninja);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Ninja criado com sucesso: " + ninjaDTO.getNome() + "(ID) : " + ninjaDTO.getId());
+
     }
 
     @GetMapping("/listar")
-    public List<NinjaDTO> listarNinjas() {
-        return ninjaService.listarNinjas();
+    public ResponseEntity<List<NinjaDTO>> listarNinjas() {
+        List<NinjaDTO> ninjas = ninjaService.listarNinjas();
+        return  ResponseEntity.ok(ninjas);
     }
 
     @GetMapping("/listar/{id}")
-    public NinjaDTO listarNinjasPorID(@PathVariable Long id) {
-        return ninjaService.buscarNinjaPorID(id);
+    public ResponseEntity<?> listarNinjasPorID(@PathVariable Long id) {
+        NinjaDTO ninja = ninjaService.buscarNinjaPorID(id);
+        if (ninjaService.buscarNinjaPorID(id) != null) {
+            return ResponseEntity.ok(ninja);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Não encontrado de ID " + id);
+        }
     }
 
     @PutMapping("/alterar/{id}")
-    public NinjaDTO alterarNinjaPorID(@PathVariable Long id, @RequestBody NinjaDTO ninja) {
-        return ninjaService.atualizarNinja(id, ninja);
+    public ResponseEntity<String> alterarNinjaPorID(@PathVariable Long id, @RequestBody NinjaDTO ninja) {
+        if (ninjaService.buscarNinjaPorID(id) != null) {
+            ninjaService.atualizarNinja(id, ninja);
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body("Ninja de ID " + id + " alterado com sucesso!");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("ID: " + id + " Não encontrado!");
+        }
     }
 
     @DeleteMapping("/deletar/{id}")
-    public void deletarNinjaPorID(@PathVariable Long id) {
-        ninjaService.deleteNinja(id);
+    public ResponseEntity<String> deletarNinjaPorID(@PathVariable Long id) {
+        if (ninjaService.buscarNinjaPorID(id) != null) {
+            ninjaService.deleteNinja(id);
+            return ResponseEntity.ok("iD: " + id + " Ninja atualizada com sucesso!");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("ID: " + id + " Não encontrado!");
+        }
     }
-
 }
